Refactor blending module into main module and submodule with blending algorithms
