+++ src/blending/algorithms.rs
@@ -1,119 +1,8 @@
-use super::utils::{max, min};
+use crate::utils::{max, min};
 
-use image::{ImageBuffer, Rgba};
-use std::fmt;
-use std::fmt::{Display, Formatter};
-use std::result;
-use std::str::FromStr;
+use image::Rgba;
 
-pub enum BlendAlgorithm {
-    Alpha,
-    Multiplicative,
-    SourceOver,
-    DestinationOver,
-    FirstTop,
-    FirstBottom,
-    DisjointOver,
-    DisjointUnder,
-    DisjointDebug,
-}
-
-impl FromStr for BlendAlgorithm {
-    type Err = String;
-
-    fn from_str(s: &str) -> result::Result<Self, Self::Err> {
-        match s {
-            "alpha" => Ok(BlendAlgorithm::Alpha),
-            "multiplicative" => Ok(BlendAlgorithm::Multiplicative),
-            "source_over" => Ok(BlendAlgorithm::SourceOver),
-            "destination_over" => Ok(BlendAlgorithm::DestinationOver),
-            "first_top" => Ok(BlendAlgorithm::FirstTop),
-            "first_bottom" => Ok(BlendAlgorithm::FirstBottom),
-            "disjoint_over" => Ok(BlendAlgorithm::DisjointOver),
-            "disjoint_under" => Ok(BlendAlgorithm::DisjointUnder),
-            "disjoint_debug" => Ok(BlendAlgorithm::DisjointDebug),
-            _ => Err(format!("'{}' is not a valid value for BlendAlgorithm", s)),
-        }
-    }
-}
-
-impl BlendAlgorithm {
-    pub fn all() -> Vec<&'static str> {
-        vec![
-            "alpha",
-            "multiplicative",
-            "source_over",
-            "destination_over",
-            "first_top",
-            "first_bottom",
-            "disjoint_over",
-            "disjoint_under",
-            "disjoint_debug",
-        ]
-    }
-}
-
-impl Display for BlendAlgorithm {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match self {
-            BlendAlgorithm::Alpha => write!(f, "alpha"),
-            BlendAlgorithm::Multiplicative => write!(f, "multiplicative"),
-            BlendAlgorithm::SourceOver => write!(f, "source_over"),
-            BlendAlgorithm::DestinationOver => write!(f, "destination_over"),
-            BlendAlgorithm::FirstTop => write!(f, "first_top"),
-            BlendAlgorithm::FirstBottom => write!(f, "first_bottom"),
-            BlendAlgorithm::DisjointOver => write!(f, "disjoint_over"),
-            BlendAlgorithm::DisjointUnder => write!(f, "disjoint_under"),
-            BlendAlgorithm::DisjointDebug => write!(f, "disjoint_debug"),
-        }
-    }
-}
-
-pub enum Background {
-    Alpha,
-    White,
-    Blue,
-    Texture,
-}
-
-impl Display for Background {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match self {
-            Background::Alpha => write!(f, "alpha"),
-            Background::White => write!(f, "white"),
-            Background::Blue => write!(f, "blue"),
-            Background::Texture => write!(f, "texture"),
-        }
-    }
-}
-
-pub fn get_blending_algorithm(
-    algorithm: &BlendAlgorithm,
-) -> impl Fn((&mut Rgba<u8>, &Rgba<u8>)) -> () {
-    match algorithm {
-        BlendAlgorithm::Alpha => blend_alpha,
-        BlendAlgorithm::Multiplicative => blend_multiplicative,
-        BlendAlgorithm::SourceOver => blend_source_over,
-        BlendAlgorithm::DestinationOver => blend_destination_over,
-        BlendAlgorithm::FirstTop => blend_first_top,
-        BlendAlgorithm::FirstBottom => blend_first_bottom,
-        BlendAlgorithm::DisjointOver => blend_disjoint_over,
-        BlendAlgorithm::DisjointUnder => blend_disjoint_under,
-        BlendAlgorithm::DisjointDebug => blend_disjoint_debug,
-    }
-}
-
-pub fn blend_images(
-    top: &ImageBuffer<Rgba<u8>, Vec<u8>>,
-    bot: &mut ImageBuffer<Rgba<u8>, Vec<u8>>,
-    blending_algorithm: &impl Fn((&mut Rgba<u8>, &Rgba<u8>)) -> (),
-) {
-    for pixel_pair in bot.pixels_mut().zip(top.pixels()) {
-        blending_algorithm(pixel_pair);
-    }
-}
-
-fn blend_alpha((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_alpha((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -148,7 +37,7 @@ fn blend_alpha((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a as u8;
 }
 
-fn blend_multiplicative((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_multiplicative((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -169,7 +58,7 @@ fn blend_multiplicative((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a as u8;
 }
 
-fn blend_source_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_source_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -204,7 +93,7 @@ fn blend_source_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a as u8;
 }
 
-fn blend_destination_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_destination_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -239,7 +128,7 @@ fn blend_destination_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a as u8;
 }
 
-fn blend_first_top((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_first_top((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -259,7 +148,7 @@ fn blend_first_top((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a;
 }
 
-fn blend_first_bottom((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_first_bottom((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -280,7 +169,7 @@ fn blend_first_bottom((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
 }
 
 //TODO: debug different result from original pconvert
-fn blend_disjoint_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_disjoint_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -314,7 +203,7 @@ fn blend_disjoint_over((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a as u8;
 }
 
-fn blend_disjoint_under((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_disjoint_under((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let (rb, gb, bb, ab) = (bot_pixel[0], bot_pixel[1], bot_pixel[2], bot_pixel[3]);
     let (rt, gt, bt, at) = (top_pixel[0], top_pixel[1], top_pixel[2], top_pixel[3]);
 
@@ -348,7 +237,7 @@ fn blend_disjoint_under((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[3] = a as u8;
 }
 
-fn blend_disjoint_debug((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
+pub fn blend_disjoint_debug((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     let ab = bot_pixel[3];
     let at = top_pixel[3];
 
@@ -368,56 +257,4 @@ fn blend_disjoint_debug((bot_pixel, top_pixel): (&mut Rgba<u8>, &Rgba<u8>)) {
     bot_pixel[1] = g;
     bot_pixel[2] = b;
     bot_pixel[3] = a as u8;
-}
-
-pub fn demultiply_image(img: &mut ImageBuffer<Rgba<u8>, Vec<u8>>) {
-    for pixel in img.pixels_mut() {
-        demultiply_pixel(pixel);
-    }
-}
-
-fn demultiply_pixel(pixel: &mut Rgba<u8>) {
-    let (r, g, b, a) = (pixel[0], pixel[1], pixel[2], pixel[3]);
-    let af = a as f32 / 255.0;
-
-    let r = (r as f32 * af).round() as u8;
-    let g = (g as f32 * af).round() as u8;
-    let b = (b as f32 * af).round() as u8;
-
-    pixel[0] = r;
-    pixel[1] = g;
-    pixel[2] = b;
-}
-
-pub fn multiply_image(img: &mut ImageBuffer<Rgba<u8>, Vec<u8>>) {
-    for pixel in img.pixels_mut() {
-        multiply_pixel(pixel);
-    }
-}
-
-fn multiply_pixel(pixel: &mut Rgba<u8>) {
-    let (r, g, b, a) = (pixel[0], pixel[1], pixel[2], pixel[3]);
-    let af = a as f32 / 255.0;
-
-    let r = (r as f32 / af).round() as u8;
-    let g = (g as f32 / af).round() as u8;
-    let b = (b as f32 / af).round() as u8;
-
-    pixel[0] = r;
-    pixel[1] = g;
-    pixel[2] = b;
-}
-
-pub fn is_algorithm_multiplied(algorithm: &BlendAlgorithm) -> bool {
-    match algorithm {
-        BlendAlgorithm::Alpha => false,
-        BlendAlgorithm::Multiplicative => false,
-        BlendAlgorithm::SourceOver => false,
-        BlendAlgorithm::DestinationOver => false,
-        BlendAlgorithm::FirstTop => false,
-        BlendAlgorithm::FirstBottom => false,
-        BlendAlgorithm::DisjointOver => true,
-        BlendAlgorithm::DisjointUnder => true,
-        BlendAlgorithm::DisjointDebug => true,
-    }
 }
